Refactor AuthProvider interface and authentication/authorization mechanism
