+++ components/global/core/io.cellery.observability.auth/src/main/java/io/cellery/observability/auth/DcrProvider.java
@@ -23,45 +23,36 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.cellery.observability.auth.exception.AuthProviderException;
-import org.apache.commons.codec.binary.Base64;
 import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
 import org.apache.http.ParseException;
 import org.apache.http.client.HttpClient;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
-import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.util.EntityUtils;
 import org.apache.log4j.Logger;
 import org.wso2.carbon.config.ConfigurationException;
 
 import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
- * Manager for managing Open ID Connect related functionality.
+ * Dynamic Client Registration Provider for Cellery Observability.
  */
-public class AuthenticationProvider {
-    private static final Logger logger = Logger.getLogger(AuthenticationProvider.class);
+public class DcrProvider {
+    private static final Logger logger = Logger.getLogger(DcrProvider.class);
 
     private static final Gson gson = new Gson();
     private static final JsonParser jsonParser = new JsonParser();
+
     private static final String ERROR = "error";
-    private static final String ACTIVE_STATUS = "active";
-    private static final String BASIC_AUTH = "Basic ";
 
     private String clientId;
     private char[] clientSecret;
 
-    public AuthenticationProvider() {
+    public DcrProvider() {
         try {
             retrieveClientCredentials();
         } catch (AuthProviderException e) {
@@ -108,7 +99,7 @@ public class AuthenticationProvider {
      */
     private void retrieveClientCredentials() throws AuthProviderException {
         if (this.clientId == null || this.clientSecret == null) {
-            JsonObject jsonObject = createClientWithDcr();
+            JsonObject jsonObject = createNewClient();
             if (jsonObject.has(ERROR)) {
                 logger.info("Fetching the credentials of the already existing client "
                         + Constants.CELLERY_APPLICATION_NAME);
@@ -126,7 +117,7 @@ public class AuthenticationProvider {
      * @return the HttpPost object
      * @throws AuthProviderException if the HTTP call for creating the Client fails
      */
-    private JsonObject createClientWithDcr() throws AuthProviderException {
+    private JsonObject createNewClient() throws AuthProviderException {
         try {
             JsonArray callbackUris = new JsonArray(1);
             callbackUris.add(AuthConfig.getInstance().getCallbackUrl());
@@ -144,7 +135,8 @@ public class AuthenticationProvider {
             StringEntity requestEntity = new StringEntity(gson.toJson(clientJson), ContentType.APPLICATION_JSON);
 
             HttpPost request = new HttpPost(dcrEp);
-            request.setHeader(Constants.HEADER_AUTHORIZATION, getEncodedIdpAdminCredentials());
+            request.setHeader(Constants.HEADER_AUTHORIZATION, Utils.generateBasicAuthHeaderValue(
+                    AuthConfig.getInstance().getIdpUsername(), AuthConfig.getInstance().getIdpPassword()));
             request.setHeader(Constants.HEADER_CONTENT_TYPE, Constants.CONTENT_TYPE_APPLICATION_JSON);
             request.setEntity(requestEntity);
 
@@ -171,7 +163,8 @@ public class AuthenticationProvider {
             String dcrEp = AuthConfig.getInstance().getIdpUrl() + Constants.OIDC_REGISTER_ENDPOINT;
             HttpGet request = new HttpGet(dcrEp + "?"
                     + Constants.OIDC_CLIENT_NAME_KEY + "=" + Constants.CELLERY_APPLICATION_NAME);
-            request.setHeader(Constants.HEADER_AUTHORIZATION, getEncodedIdpAdminCredentials());
+            request.setHeader(Constants.HEADER_AUTHORIZATION, Utils.generateBasicAuthHeaderValue(
+                    AuthConfig.getInstance().getIdpUsername(), AuthConfig.getInstance().getIdpPassword()));
 
             HttpClient client = Utils.getTrustAllClient();
             HttpResponse response = client.execute(request);
@@ -188,55 +181,5 @@ public class AuthenticationProvider {
             throw new AuthProviderException("Error occurred while retrieving the client credentials with name " +
                     Constants.CELLERY_APPLICATION_NAME, e);
         }
-    }
-
-    /**
-     * Validate the token.
-     *
-     * @param token This will be the access token
-     * @return the boolean status of the validity of the access token
-     * @throws AuthProviderException if validating the token fails
-     */
-    public boolean validateToken(String token) throws AuthProviderException {
-        try {
-            List<NameValuePair> params = new ArrayList<>();
-            params.add(new BasicNameValuePair("token", token));
-
-            String introspectEP = AuthConfig.getInstance().getIdpUrl() + Constants.OIDC_INTROSPECT_ENDPOINT;
-            HttpPost request = new HttpPost(introspectEP);
-            request.setHeader(Constants.HEADER_AUTHORIZATION, getEncodedIdpAdminCredentials());
-            request.setEntity(new UrlEncodedFormEntity(params, StandardCharsets.UTF_8.name()));
-
-            HttpClient client = Utils.getTrustAllClient();
-            HttpResponse response = client.execute(request);
-            int statusCode = response.getStatusLine().getStatusCode();
-            if (statusCode >= 200 && statusCode < 400) {
-                JsonObject jsonObject = jsonParser.parse(EntityUtils.toString(response.getEntity())).getAsJsonObject();
-                if (!jsonObject.get(ACTIVE_STATUS).getAsBoolean()) {
-                    return false;
-                }
-            } else {
-                logger.error("Failed to connect to Introspect endpoint in Identity Provider server." +
-                        " Exited with Status Code " + statusCode);
-                return false;
-            }
-        } catch (IOException | ParseException | NoSuchAlgorithmException | KeyManagementException |
-                ConfigurationException e) {
-            throw new AuthProviderException("Error occurred while calling the introspect endpoint", e);
-        }
-        return true;
-    }
-
-    /**
-     * Get the Base64 encoded IdP Admin Credentials.
-     *
-     * @return the String value of encoded credential is returned
-     */
-    private String getEncodedIdpAdminCredentials() throws ConfigurationException {
-        String authString = AuthConfig.getInstance().getIdpUsername() + ":"
-                + AuthConfig.getInstance().getIdpPassword();
-        byte[] authEncBytes = Base64.encodeBase64(authString.getBytes(Charset.forName(StandardCharsets.UTF_8.name())));
-        String authStringEnc = new String(authEncBytes, Charset.forName(StandardCharsets.UTF_8.name()));
-        return BASIC_AUTH + authStringEnc;
     }
 }
